separated UiGrid class
